requirements.txt
boto3>=1.34.0
anthropic>=0.39.0
streamlit>=1.32.0
pandas>=2.2.0
python-dotenv>=1.0.0
